create our emitter_v2
Our emitter_v2 generate xml in the format we want and with the attributes we care.
It also dumpy numpy file for each const and check their value with the binary file containing all the const value.
Creating our own emitter also allows us to choose which emitter(openvino's or ours) to use to emit ir depending on our purpose.
Our emitter is based on the original one in
5cee8bbf2 (tag: 2021.4) omg layout fix (#6450)