this pass does type inference for the model graph
When implementing Op IRs generating parts, we observe that if type inference are done in emitter part codes, it causes error.
So the purpose of this pass is to replace the type_infer function call in emitter part codes.
It has the same functionality as the function call, but causes no error when emitting op IRs.